Modified PfSense to add custom functionality to update DNS.
The functionality was added as follows : Custom provider in Pfsense
files, similar to the new functionality existing in versions 2.1+ where
you can use your own custom provider.

Along with it, a PHP SSHSecLib and a script to capture the new IP with a
basic security. This was a try out model... works perfectly fine.